Move etcd_util.go to separate package
