+++ pkg/storage/etcd/etcd_helper_test.go
@@ -17,19 +17,11 @@ limitations under the License.
 package etcd
 
 import (
-	"fmt"
-	"math/rand"
-	"net"
-	"net/http"
-	"net/http/httptest"
 	"path"
 	"reflect"
-	"strconv"
 	"sync"
 	"testing"
-	"time"
 
-	"github.com/coreos/go-etcd/etcd"
 	"github.com/stretchr/testify/assert"
 	"golang.org/x/net/context"
 	"k8s.io/kubernetes/pkg/api"
@@ -40,6 +32,7 @@ import (
 	"k8s.io/kubernetes/pkg/storage"
 	"k8s.io/kubernetes/pkg/storage/etcd/etcdtest"
 	etcdtesting "k8s.io/kubernetes/pkg/storage/etcd/testing"
+	etcdutil "k8s.io/kubernetes/pkg/storage/etcd/util"
 	storagetesting "k8s.io/kubernetes/pkg/storage/testing"
 
 	// TODO: once fakeClient has been purged move utils
@@ -67,18 +60,6 @@ func init() {
 
 func newEtcdHelper(client tools.EtcdClient, codec runtime.Codec, prefix string) etcdHelper {
 	return *NewEtcdStorage(client, codec, prefix).(*etcdHelper)
-}
-
-func TestIsEtcdNotFound(t *testing.T) {
-	try := func(err error, isNotFound bool) {
-		if IsEtcdNotFound(err) != isNotFound {
-			t.Errorf("Expected %#v to return %v, but it did not", err, isNotFound)
-		}
-	}
-	try(tools.EtcdErrorNotFound, true)
-	try(&etcd.EtcdError{ErrorCode: 101}, false)
-	try(nil, false)
-	try(fmt.Errorf("some other kind of error"), false)
 }
 
 // Returns an encoded version of api.Pod with the given name.
@@ -265,7 +246,7 @@ func TestGetNotFoundErr(t *testing.T) {
 
 	var got api.Pod
 	err := helper.Get(context.TODO(), boguskey, &got, false)
-	if !IsEtcdNotFound(err) {
+	if !etcdutil.IsEtcdNotFound(err) {
 		t.Errorf("Unexpected reponse on key=%v, err=%v", key, err)
 	}
 }
@@ -539,53 +520,6 @@ func TestGuaranteedUpdate_CreateCollision(t *testing.T) {
 	}
 }
 
-func TestGetEtcdVersion_ValidVersion(t *testing.T) {
-	testServer := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		fmt.Fprint(w, validEtcdVersion)
-	}))
-	defer testServer.Close()
-
-	var version string
-	var err error
-	if version, err = GetEtcdVersion(testServer.URL); err != nil {
-		t.Errorf("Unexpected error: %v", err)
-	}
-	assert.Equal(t, validEtcdVersion, version, "Unexpected version")
-	assert.Nil(t, err)
-}
-
-func TestGetEtcdVersion_ErrorStatus(t *testing.T) {
-	testServer := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.WriteHeader(http.StatusServiceUnavailable)
-	}))
-	defer testServer.Close()
-
-	_, err := GetEtcdVersion(testServer.URL)
-	assert.NotNil(t, err)
-}
-
-func TestGetEtcdVersion_NotListening(t *testing.T) {
-	portIsOpen := func(port int) bool {
-		conn, err := net.DialTimeout("tcp", "127.0.0.1:"+strconv.Itoa(port), 1*time.Second)
-		if err == nil {
-			conn.Close()
-			return true
-		}
-		return false
-	}
-
-	port := rand.Intn((1 << 16) - 1)
-	for tried := 0; portIsOpen(port); tried++ {
-		if tried >= 10 {
-			t.Fatal("Couldn't find a closed TCP port to continue testing")
-		}
-		port++
-	}
-
-	_, err := GetEtcdVersion("http://127.0.0.1:" + strconv.Itoa(port))
-	assert.NotNil(t, err)
-}
-
 func TestPrefixEtcdKey(t *testing.T) {
 	server := etcdtesting.NewEtcdTestClientServer(t)
 	defer server.Terminate(t)
@@ -605,33 +539,4 @@ func TestPrefixEtcdKey(t *testing.T) {
 	keyAfter = helper.prefixEtcdKey(keyBefore)
 
 	assert.Equal(t, keyBefore, keyAfter, "Prefix incorrectly added by EtcdHelper")
-}
-
-func TestEtcdHealthCheck(t *testing.T) {
-	tests := []struct {
-		data      string
-		expectErr bool
-	}{
-		{
-			data:      "{\"health\": \"true\"}",
-			expectErr: false,
-		},
-		{
-			data:      "{\"health\": \"false\"}",
-			expectErr: true,
-		},
-		{
-			data:      "invalid json",
-			expectErr: true,
-		},
-	}
-	for _, test := range tests {
-		err := EtcdHealthCheck([]byte(test.data))
-		if err != nil && !test.expectErr {
-			t.Errorf("unexpected error: %v", err)
-		}
-		if err == nil && test.expectErr {
-			t.Error("unexpected non-error")
-		}
-	}
 }
